+++ math/src/dsle.c
@@ -6,7 +6,7 @@
 #include <mtx.h>
 #include <sle.h>
 
-static inline void swap(size_t* p, size_t a, size_t b) {
+static inline void swap(int* p, int a, int b) {
   if (a == b)
     return;
 
@@ -16,16 +16,16 @@ static inline void swap(size_t* p, size_t a, size_t b) {
 }
 
 void sle_gauss(struct mtx* mp, struct vec* xp, struct vec* bp) {
-  size_t* p = malloc(sizeof(size_t) * mp->n);
+  int* p = malloc(sizeof(int) * mp->n);
 
-  for (size_t i = 0; i < mp->n; ++i)
+  for (int i = 0; i < mp->n; ++i)
     p[i] = i;
 
-  for (size_t i = 0; i < mp->n; ++i) {
+  for (int i = 0; i < mp->n; ++i) {
     real m = fabs(ME(mp, p[i], i));
-    size_t mi = i;
+    int mi = i;
 
-    for (size_t j = i + 1; j < mp->n; ++j)
+    for (int j = i + 1; j < mp->n; ++j)
       if (fabs(ME(mp, p[j], i)) > m) {
         m = fabs(ME(mp, p[j], i));
         mi = j;
@@ -33,22 +33,22 @@ void sle_gauss(struct mtx* mp, struct vec* xp, struct vec* bp) {
 
     swap(p, i, mi);
 
-    for (size_t j = i + 1; j < mp->n; ++j) {
+    for (int j = i + 1; j < mp->n; ++j) {
       real k = ME(mp, p[j], i) / ME(mp, p[i], i);
 
-      for (size_t c = i + 1; c < mp->n; ++c)
+      for (int c = i + 1; c < mp->n; ++c)
         ME(mp, p[j], c) -= ME(mp, p[i], c) * k;
 
       bp->v[p[j]] -= bp->v[p[i]] * k;
     }
   }
 
-  for (size_t o = 0, i; o < mp->n; ++o) {
+  for (int o = 0, i; o < mp->n; ++o) {
     i = mp->n - 1 - o;
 
     xp->v[i] = bp->v[p[i]];
 
-    for (size_t j = i + 1; j < mp->n; ++j)
+    for (int j = i + 1; j < mp->n; ++j)
       xp->v[i] -= xp->v[j] * ME(mp, p[i], j);
 
     xp->v[i] /= ME(mp, p[i], i);
